+++ Source/Views/Tire Set/TireSetDetailsView.swift
@@ -26,11 +26,11 @@ struct TireSetDetailsView: View {
     var body: some View {
         FleetboxTextField(value: $tireSet.userDisplayName, name: "Name", example: "My Summer Tires")
         FleetboxTextField(value: $tireSet.make, name: "Make", example: "TireCo")
-            .textInputAutocapitalization(.words)
+            .autocapitalization(.words)
         FleetboxTextField(value: $tireSet.model, name: "Model", example: "Aviator Sport")
-            .textInputAutocapitalization(.words)
+            .autocapitalization(.words)
         FleetboxTextField(value: $tireSet.tin, name: "TIN", example: "DOT U2LL LMLR5107")
-            .textInputAutocapitalization(.characters)
+            .autocapitalization(.characters)
         if editable {
             if tireSet.hidden {
                 Button("Un-hide") {