+++ Source/Views/Vehicle/VINDetailView.swift
@@ -60,7 +60,7 @@ struct VINDetailView: View {
     var body: some View {
         HStack {
             FleetboxTextField(value: $vehicle.vin, name: "VIN", example: dummyData.vin)
-                .textInputAutocapitalization(.characters)
+                .autocapitalization(.characters)
             if editable, let vin = vehicle.vin, vin != "" {
                 switch state {
                 case .base: